refactor(surfaces): extract shared tool-label formatting into surfaces/shared (T-21)
surfaces/cli/ui/renderer/tools.py and surfaces/interactive_shell/ui/output/
tool_details.py each defined identical tool_source_label/infer_tool_source/
tool_short_label functions for rendering live tool-call activity badges.
Pure formatting with no surface-specific state, so it's the justified case
T-21 asks for: extracted into surfaces/shared/tool_labels.py, with both
surfaces (and interactive_shell's tool_tracking.py and
integration_tool_gathering.py, which also used the interactive_shell copy)
now importing the single implementation.

No other genuine cross-surface duplication was found on this pass; other
tools/ and ui/ modules in each surface diverge in real ways (rich Text
rendering, REPL-specific state) rather than being copy-pasted.
