+++ surfaces/cli/ui/renderer/tools.py
@@ -1,10 +1,13 @@
-"""Tool-call labeling and payload extraction helpers."""
+"""Tool-call event-payload extraction helpers.
+
+Source/label formatting (``tool_source_label``, ``tool_short_label``) moved
+to ``surfaces.shared.tool_labels`` (T-21) — it was identical to the
+interactive shell's copy.
+"""
 
 from __future__ import annotations
 
 from typing import Any
-
-from tools.registry import get_registered_tool_map, resolve_tool_display_name
 
 
 def _tool_event_key(data: dict[str, Any], name: str) -> str:
@@ -16,46 +19,6 @@ def _tool_event_key(data: dict[str, Any], name: str) -> str:
         or nested.get("tool_call_id")
         or name
     )
-
-
-def _tool_source_label(tool_name: str) -> str:
-    tool = get_registered_tool_map().get(tool_name)
-    source = str(tool.source) if tool is not None else _infer_tool_source(tool_name)
-    if source == "grafana":
-        return "Grafana"
-    if source == "knowledge":
-        return "SRE"
-    if source == "openclaw":
-        return "OpenClaw"
-    return source.replace("_", " ").title() if source else "Tools"
-
-
-def _infer_tool_source(tool_name: str) -> str:
-    lowered = tool_name.lower()
-    for source in ("grafana", "datadog", "cloudwatch", "sentry", "honeycomb", "openclaw"):
-        if source in lowered:
-            return source
-    if lowered.startswith("get_sre_"):
-        return "knowledge"
-    return "tools"
-
-
-def _tool_short_label(tool_name: str, source_label: str) -> str:
-    display = resolve_tool_display_name(tool_name)
-    label = display
-    for prefix in (
-        source_label,
-        source_label.lower(),
-        f"{source_label} ",
-        f"{source_label.lower()} ",
-        "query ",
-        "get ",
-    ):
-        if label.startswith(prefix):
-            label = label[len(prefix) :].strip()
-    if source_label == "Grafana" and label.lower().startswith("grafana "):
-        label = label[len("grafana ") :].strip()
-    return label or display
 
 
 def _tool_input(data: dict[str, Any]) -> Any:
